+++ logic/gpg.go
@@ -42,7 +42,7 @@ func (t *Transaction) execAddGPGKey(
 
 	// Get sender accounts
 	sender := spk.Addr()
-	senderAcct := acctKeeper.GetAccount(sender, int64(chainHeight))
+	senderAcct := acctKeeper.GetAccount(sender, chainHeight)
 	senderBal := senderAcct.Balance.Decimal()
 
 	// Deduct the fee from the sender's account
